Move all utils to lib/
