+++ labextension/src/lib/CellUtils.ts
