+++ labextension/src/lib/NotebookUtils.tsx
